añadida categorias a la base + dao

+++ src/main/java/persistence/ProposalDao.java
@@ -4,6 +4,6 @@ import dto.Proposal;
 
 public interface ProposalDao {
 
-	Proposal getProposalById(Long id);
+	Proposal getProposalById(Integer id);
 
 }
